Fix config listener event handler registration
We rely on `gen_event:add_sup_handler/3` to remove handlers when the
process that registered for events exits. On master this was changed so
that config becomes the process that's monitored by gen_event. As such
any handler that is registered (say, for when an index is opened) adds a
handler to the config_event gen_event process. Since the config process
never exits these handlers are never removed.

The end result of all of this is that on a busy cluster the config_event
process will end up with millions of handlers consuming many gigabytes
of RAM.

This change creates a monitor process for every event handler. This
monitors the process wanting to listen for config changes and exits when
the requesting process exits. This means that we maintain our pure
callback API improvement while correctly removing handlers.

COUCHDB-3096