Clean up config_tests
This cleans up the config_tests both stylistically as well as removes
some race conditions around message passing from the config handler.

This also reformats and changes a lot of the tests so that we're using
consistent patterns through out the file. Unfortunately foreach and
foreachx are terrible constructs and require using the `?_test/1` macro
which is a bit annoying but ended up being the least worst approach I
could find.

COUCHDB-3096